# net,net/netip: implement the encoding.(Binary|Text)Appender