--- a/doc/next/6-stdlib/99-minor/net/62384.md
+++ b/doc/next/6-stdlib/99-minor/net/62384.md
@@ -1,0 +1,1 @@
+[IP] now implements the [encoding.TextAppender] interface.