--- a/doc/next/6-stdlib/99-minor/net/netip/62384.md
+++ b/doc/next/6-stdlib/99-minor/net/netip/62384.md
@@ -1,0 +1,2 @@
+[Addr], [AddrPort] and [Prefix] now implement the [encoding.BinaryAppender] and
+[encoding.TextAppender] interfaces.